+++ src/basics/Constants.sol
@@ -3,10 +3,10 @@ pragma solidity 0.8.19;
 
 contract Constants {
     address public constant MY_ADDRESS = address(0);
-    uint public constant MY_UINT = 134131;
+    uint256 public constant MY_UINT = 134131;
 }
 
 contract Vars {
     address public MY_ADDRESS = address(0);
-    uint public MY_UINT = 134131;
+    uint256 public MY_UINT = 134131;
 }
