+++ src/basics/GlobalVars.sol
@@ -3,12 +3,10 @@ pragma solidity 0.8.19;
 
 contract GlobalVariables {
     // View because you are reading global variables
-    function globalVars() external view returns(address, uint, uint) {
+    function globalVars() external view returns (address, uint256, uint256) {
         address sender = msg.sender;
-        uint timestamp = block.timestamp;
-        uint number = block.number;
+        uint256 timestamp = block.timestamp;
+        uint256 number = block.number;
         return (sender, timestamp, number);
     }
-
 }
-
